+++ src/main/java/com/example/surveyapp/domain/surveyanswer/domain/strategy/MultiChoiceAnswerStrategy.java
@@ -1,13 +1,12 @@
 package com.example.surveyapp.domain.surveyanswer.domain.strategy;
 
-import com.example.surveyapp.domain.surveyanswer.controller.dto.request.QuestionAnswerRequestDto;
-import com.example.surveyapp.domain.survey.domain.model.entity.Question;
+import com.example.surveyapp.domain.surveyanswer.exception.AnswerErrorCode;
+import com.example.surveyapp.domain.surveyanswer.exception.AnswerException;
+import com.example.surveyapp.domain.surveyanswer.presentation.dto.request.QuestionAnswerRequestDto;
 import com.example.surveyapp.domain.surveyanswer.domain.model.SurveyAnswer;
 import com.example.surveyapp.domain.surveyanswer.domain.model.SurveyOptionsAnswer;
 import com.example.surveyapp.domain.survey.domain.model.enums.QuestionType;
 import com.example.surveyapp.domain.surveyanswer.domain.repository.SurveyOptionsAnswerRepository;
-import com.example.surveyapp.global.response.exception.CustomException;
-import com.example.surveyapp.global.response.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -26,7 +25,7 @@ public class MultiChoiceAnswerStrategy implements SurveyQuestionStrategy {
                 Long number = Long.parseLong(s);
                 surveyOptionsAnswerRepository.save(new SurveyOptionsAnswer(surveyAnswer.getId(), questionId, number));
             } catch (NumberFormatException e) {
-                throw new CustomException(ErrorCode.VALIDATION_ERROR);
+                throw new AnswerException(AnswerErrorCode.VALIDATION_ERROR);
             }
         }
     }
