+++ src/main/java/com/example/surveyapp/domain/surveyanswer/domain/strategy/SingleChoiceAnswerStrategy.java
@@ -1,7 +1,6 @@
 package com.example.surveyapp.domain.surveyanswer.domain.strategy;
 
-import com.example.surveyapp.domain.surveyanswer.controller.dto.request.QuestionAnswerRequestDto;
-import com.example.surveyapp.domain.survey.domain.model.entity.Question;
+import com.example.surveyapp.domain.surveyanswer.presentation.dto.request.QuestionAnswerRequestDto;
 import com.example.surveyapp.domain.surveyanswer.domain.model.SurveyAnswer;
 import com.example.surveyapp.domain.surveyanswer.domain.model.SurveyOptionsAnswer;
 import com.example.surveyapp.domain.survey.domain.model.enums.QuestionType;
