+++ src/main/java/com/example/surveyapp/domain/surveyanswer/domain/strategy/SurveyQuestionStrategy.java
@@ -1,6 +1,6 @@
 package com.example.surveyapp.domain.surveyanswer.domain.strategy;
 
-import com.example.surveyapp.domain.surveyanswer.controller.dto.request.QuestionAnswerRequestDto;
+import com.example.surveyapp.domain.surveyanswer.presentation.dto.request.QuestionAnswerRequestDto;
 import com.example.surveyapp.domain.surveyanswer.domain.model.SurveyAnswer;
 import com.example.surveyapp.domain.survey.domain.model.enums.QuestionType;
